feat: Use selenized-dark color palette
Websites often provide a "dark" mode for low-light settings. However, it
is a burden to add features while supporting "light" and "dark" mode.

Instead, use a color palette that is designed to be used in both low and
high light settings.

Selenized is a modified version of the Solarized color palette.