fix: Change <div class="header"> to <h1> and <p>
A block element is not needed to get the desired visual effect

Semantically, a <div> was not appropriate - the section is not separate
from the rest of the page. It also caused unintended visual effects on
text browsers.

In the CSS, <h1> was given no bottom margin; this is to prevent margin
collapse from altering the intended style of the page. If a margin is
needed, use "margin-top" from the following element.