feat!: Add bar to navigate contact information
With the previous format, when text would overflow, it would severely
degrade the aesthetic of the page. This is especially a problem on
mobile devices which have a smaller horizontal display.

The bar was introduced to mitigate this issue. It is a common UI element
and works well with horizontal scrolling - which handles text overflow.

BREAKING CHANGE: use CSS and HTML features not supported by HTML 2.0