fix: Remove about.htm from website
The goal of "about" section was to explicitly mention my name. However,
beyond this, it served no purpose. After an attempt to turn it into
something sensible, it was deemed more appropriate to deprecate.

There is no known case where someone visiting this website is not aware
of my name. If from github (or equivalent), my name is on the profile.
If from a resume, my name is on the document. Web-scrapers are blocked,
it is unlikely to visit via search-engine.